+++ frontend-ds/gestor-proyecto/src/app/views/get_frontend/get_cursos.components.ts
@@ -1,6 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import { AppService } from 'src/app/app.service';
+import swal from 'sweetalert2';
+
 
 @Component({
     selector: 'get_frontend',
@@ -32,7 +34,11 @@ export class GetCursoComponent implements OnInit{
         this.service.get_cursos().subscribe(
             data => response = data,
             err => {
-                console.log("Error al consultar el servicio");
+                swal.fire({
+                    icon: 'error',
+                    title: 'Oops...',
+                    text: 'Algo salio mal!',
+                })
             },
             () => {
                 this.listado_curso = response;
@@ -46,12 +52,22 @@ export class GetCursoComponent implements OnInit{
         this.service.insert_curso(this.curso).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.curso={
                 desc_curso:""
             }
+            swal.fire({
+                title: 'Curso agregado exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
         this.get_cursos();
         }
         );
@@ -63,15 +79,38 @@ export class GetCursoComponent implements OnInit{
         var load={
             cod_curso:id
         }
-        this.service.delete_curso(load).subscribe(
-        data => response = data,
-        err => {
-        console.log("Error al consultar servicio");
-        },
-        ()=> {
-            this.get_cursos();
-        }
-        );
+        swal.fire({
+            title: 'Advertencia!',
+            text: "Esta seguro desea borrar este curso?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+          }).then((result) => {
+            if (result.value) {
+                this.service.delete_curso(load).subscribe(
+                    data => response = data,
+                    err => {
+                        swal.fire({
+                            icon: 'error',
+                            title: 'Oops...',
+                            text: 'Algo salio mal!',
+                        })
+                    },
+                    ()=> {
+                        this.get_cursos();
+                    }
+                    );
+              swal.fire(
+                'Borrado!',
+                'El curso a sido borrado.',
+                'success'
+              )
+            }
+          })
+        
     }
 
 
@@ -83,18 +122,25 @@ export class GetCursoComponent implements OnInit{
         var load={
             cod_curso:this.id
         };
-
-        console.log("adssadsadsadasdsadsa: " + this.curso[0]);
-
         this.service.update_curso(load, this.curso).subscribe(
         data => response = data,
         err => {
-        console.log("Error al consultar servicio");
+            swal.fire({
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Algo salio mal!',
+            })
         },
         ()=> {
             this.curso = {
                 desc_curso:""
             };
+            swal.fire({
+                title: 'Curso actualizado exitosamente!',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500
+            })
             this.get_cursos();
         }
         );
@@ -116,9 +162,4 @@ export class GetCursoComponent implements OnInit{
 
     }
 
-    toggleShow1() {
-
-        this.isShown = ! this.isShown;
-    
-        }
 }
